Extend search for cache on miss
If the exact key is not found it will first search for cache hits for the same service on previous commits.
If none are found the search is extended to all other services on that branch.
Both steps are repeated on master if no cache was hit on the branch.
Further details: https://docs.github.com/en/actions/using-workflows/caching-dependencies-to-speed-up-workflows#matching-a-cache-key