HOTFIX: Take UTC times from RCC MDF Chunks
The MDF chunks from RCC contain not only the PTS start and end times,
but also the UTC ones.
Due to the fact that RCC stores MDF chunks in blocks of one hour, a concatenated
MDF file most often represents more frames than the actual recording has.
Therefore the synchronizer in MDF Parser needs the UTC timestamps of the
MDF chunks, so that it can cut the frame to the ones of the actual recording.