Use GracefulExit handler right after startup
There are occasions where the container is shut down during waiting for the feature chain container. It looks like the backoff
mechanism used for that is run in a separate thread and therefore ignores the system.exit(1) emitted by the ImmediateExit handler.
To fix it we use the GracefulExit handler right from the start and terminate after wait_for_featurechain() is done
by relying on the existing shutdown mechanism of the loop. This could take a bit longer until the waiting finished
but should be more reliable in overall termination of the container.