Set up a dedicated sigterm handler for the API process
This fixes a nasty bug which lets the handler pod to never terminate gracefully.
When setting up a custom signal handler before forking a process, this child process will take over the same signal handler as the parent.
The parent process uses the GracefulExit handler which only sets a flag to terminate the processing loop.
This is not sufficient to terminate the Flask server. The Flask server reacts on KeyboardInterrupts therefore manually triggering that on sigterm
should terminate the server gracefully.